+++ upl/com.energyict.mdc.upl.impl/protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/nta/dsmr23/DlmsProperties.java
@@ -269,6 +269,6 @@ public class DlmsProperties implements DlmsSessionProperties {
 
     @Override
     public int getGeneralBlockTransferWindowSize() {
-        return properties.getTypedProperty(GBT_WINDOW_SIZE, DEFAULT_GBT_WINDOW_SIZE);
+        return properties.getTypedProperty(GBT_WINDOW_SIZE, DEFAULT_GBT_WINDOW_SIZE).intValue();
     }
 }
